refactor: geohash 기반 조회로 전환하게 되면서 기존 동서남북 좌표 기반 코드 정리
