+++ src/main/java/com/example/log4u/domain/diary/repository/CustomDiaryRepository.java
@@ -9,7 +9,6 @@ import com.example.log4u.domain.diary.SortType;
 import com.example.log4u.domain.diary.VisibilityType;
 import com.example.log4u.domain.diary.dto.DiaryWithAuthorDto;
 import com.example.log4u.domain.diary.entity.Diary;
-import com.example.log4u.domain.map.dto.response.DiaryMarkerResponseDto;
 
 public interface CustomDiaryRepository {
 
@@ -34,14 +33,4 @@ public interface CustomDiaryRepository {
 		Long cursorId,
 		Pageable pageable);
 
-	List<DiaryMarkerResponseDto> findDiariesInBounds(double south, double north, double west, double east);
-
-	List<Diary> findInBoundsByUserId(Long userId, double south, double north, double west, double east);
-
-	List<DiaryMarkerResponseDto> findMyDiariesInBounds(
-		Long userId,
-		double south,
-		double north,
-		double west,
-		double east);
 }
