--- src/main/java/com/example/log4u/domain/map/entity/Areas.java
@@ -1,11 +1,0 @@
-package com.example.log4u.domain.map.entity;
-
-public interface Areas {
-	String getName();
-
-	Long getId();
-
-	Double getLat();
-
-	Double getLon();
-}
