+++ src/main/java/com/example/log4u/domain/map/entity/SidoAreas.java
@@ -10,13 +10,12 @@ import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import lombok.Setter;
 
 @Entity
 @Getter
 @NoArgsConstructor
 @Table(name = "sido_areas", schema = "public")
-public class SidoAreas implements Areas {
+public class SidoAreas  {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -47,25 +46,5 @@ public class SidoAreas implements Areas {
 		this.name = name;
 		this.lat = lat;
 		this.lon = lon;
-	}
-
-	@Override
-	public String getName() {
-		return name;
-	}
-
-	@Override
-	public Long getId() {
-		return id;
-	}
-
-	@Override
-	public Double getLat() {
-		return lat;
-	}
-
-	@Override
-	public Double getLon() {
-		return lon;
 	}
 }
