+++ src/main/java/com/example/log4u/domain/map/repository/sido/SidoAreasRepository.java
@@ -1,8 +1,5 @@
 package com.example.log4u.domain.map.repository.sido;
 
-import java.util.List;
-import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
@@ -18,22 +15,6 @@ public interface SidoAreasRepository extends JpaRepository<SidoAreas, Long>, Sid
 		SELECT s FROM SidoAreas s
 		WHERE ST_Contains(s.geom, ST_SetSRID(ST_Point(:lon, :lat), 4326)) = true
 		""")
-	Optional<SidoAreas> findRegionByLatLon(@Param("lat") Double lat, @Param("lon") Double lon);
-
-	@Query("""
-		SELECT s FROM SidoAreas s
-		WHERE s.lat BETWEEN :south AND :north
-		AND s.lon BETWEEN :west AND :east
-		""")
-	List<SidoAreas> findWithinBoundingBox(@Param("south") double south, @Param("north") double north,
-		@Param("west") double west, @Param("east") double east
-	);
-
-
-	@Query("""
-		SELECT s FROM SidoAreas s
-		WHERE ST_Contains(s.geom, ST_SetSRID(ST_Point(:lon, :lat), 4326)) = true
-		""")
 	SidoAreas findSidoAreasByLatLon(@Param("lat") Double lat, @Param("lon") Double lon);
 
 	@Query("""
@@ -43,8 +24,5 @@ public interface SidoAreasRepository extends JpaRepository<SidoAreas, Long>, Sid
 		  AND ST_Contains(sNew.geom, ST_SetSRID(ST_Point(:newLon, :newLat), 4326)) = true
 		  AND sOld.id = sNew.id
 		""")
-	boolean isSameSidoRegion(@Param("oldLat") Double oldLat,
-							 @Param("oldLon") Double oldLon,
-							 @Param("newLat") Double newLat,
-							 @Param("newLon") Double newLon);
+	boolean isSameSidoRegion(@Param("oldLat") Double oldLat, @Param("oldLon") Double oldLon, @Param("newLat") Double newLat, @Param("newLon") Double newLon);
 }
