--- src/main/java/com/example/log4u/domain/map/service/strategy/SidoRegionStrategy.java
@@ -1,28 +1,0 @@
-package com.example.log4u.domain.map.service.strategy;
-
-import java.util.List;
-import java.util.Optional;
-
-import org.springframework.stereotype.Component;
-
-import com.example.log4u.domain.map.entity.SidoAreas;
-import com.example.log4u.domain.map.repository.sido.SidoAreasRepository;
-
-import lombok.RequiredArgsConstructor;
-
-@Component
-@RequiredArgsConstructor
-public class SidoRegionStrategy implements AreaRegion<SidoAreas> {
-
-	private final SidoAreasRepository sidoAreasRepository;
-
-	@Override
-	public List<SidoAreas> findRegionsInBounds(double south, double north, double west, double east) {
-		return sidoAreasRepository.findWithinBoundingBox(south, north, west, east);
-	}
-
-	@Override
-	public Optional<SidoAreas> findRegionByLatLon(double lat, double lon) {
-		return sidoAreasRepository.findRegionByLatLon(lat, lon);
-	}
-}
